requirements.txt
Flask==2.2.2
tensorflow
Pillow
transformers
torch